[Admin Panel][Insights] ConfigurationsTable (Recidiviz/recidiviz-data#27779)
## Description of the change

Adds a `ConfigurationsTable` to the Insights Confiurations view.

I also added an `ConfigurationsPresenter` that is responsible for
hydrating the configs and creating a new version of a config. This
allows the Insights Store to exist in a state where no state code is
selected, but the state code always exists in the presenter.

I don't love that the presenter is hydrating itself, but without adding
a bunch of hydration component code this seemed like the cleanest way to
do this. If we end up adding many more hydratable components, then we
spend the time to add the hydrators.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27367

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5df9cb5dfb87844f7d012676d600985c0c99476d