[Bugfix] Fix issues with pipeline integration test debug flag (Recidiviz/recidiviz-data#27857)
## Description of the change

While I was fixing test failures in Recidiviz/recidiviz-data#27811 I realized that the tests
weren't failing during the phase of the pipeline I expected. We actually
check output in two places: 1) we check the entity tree output
post-validations and 2) we check the actual table rows produced. The
tests weren't failing until the second step because the custom
`equals_to` function for entity tree lists wasn't throwing a
`BeamAssertException` (returning False does nothing). Separately, the
list comparison was actually buggy and failing with most runs because we
didn't sort the entity trees.

This PR fixes the assertion check so that it properly fails when the
lists do not contain the same elements.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Follow-up to Recidiviz/recidiviz-data#27811

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e179341bac5cb409d9fe1d6c46e43f9357fa315c