[Admin Panel][Insights] Implement /promote/default (Recidiviz/recidiviz-data#27838)
## Description of the change

Implements an endpoint for the FE to call to make any configuration with
a feature variant the new, active, default configuration.

An alternate implementation would have been to just call the existing
POST /configurations endpoint. However, I think that the FE should be
'oblivious' to the BE logic, where we're actually adding a new
configuration and deactivating the existing default configuration.

Also renames the existing /promote endpoint path to be `/promote/prod`.
An alternate name for this endpoint might be
'/configurations/<id>/set_default` if we want to decouple this action
with the idea of promoting to prod.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27535

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: fc9cf1e5bac29588115c00b4795adb36eaac203d