[Workflows] include people only eligible in new query in CR referral record (Recidiviz/recidiviz-data#27936)
## Description of the change

Previously, we LEFT JOINed the old/new CR queries together because the
frontend didn't know how to handle all the new fields. Now that it does,
we can start including people who only show up in the new style query by
doing a FULL OUTER JOIN.

If my queries are correct, this means we will start showing an
additional 214 people as eligible or almost eligible (in staging).

A different option would be to stop excluding the old records
completely, but the implications of that not working are that we
possibly show nobody as eligible, whereas the implications of this not
working are we get a bunch of sentry errors. Maybe we do that next week
after this is deployed in staging if everything goes well.

Loaded to sandbox at `dana_cr`

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes https://github.com/Recidiviz/recidiviz-data/issues/22521

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6da99617953c6faabd0e53bebcbe100b54da7ab7