[Insights] Use an empty string as a default value for two config fields (Recidiviz/recidiviz-data#27948)
## Description of the change

Updates the copy configs for US_ID and US_PA such that
exclusionReasonDescription and descriptionMarkdown are empty strings and
not null for these states. Enabling empty strings to test PA and IX in
the webtool, but the description markdown should be populated correctly
prior to beginning UXR via `outliers_configs.py` and the exclusion
reason description should be populated via the admin panel config. These
steps can happen with Recidiviz/recidiviz-data#27170

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27945

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 846ff9c619b547ce846c87997f09bbef4e4f261d