[IID Cleanup] Remove unnecessary pipeline exemption mechanisms (Recidiviz/recidiviz-data#27811)
## Description of the change

When we were building IID, we built in a few different ways to exempt
errors and slightly change logic in a way that helped with validation
but didn't necessarily impact correctness of the data. This PR deletes
those mechanisms:
1) `ENTITY_UNIQUE_CONSTRAINT_ERROR_EXEMPTIONS` - allowed us to exempt
certain entity uniqueness constraints that were failing, but we now
don't have any more of these errors, so fine to delete!
2) `GLOBAL_UNIQUE_CONSTRAINT_ERROR_EXEMPTIONS` - Same as above, but for
global unique constraints (constraints checked across all entities, not
just the ones in a root entity tree)
3) `INGEST_VIEW_ORDER_EXEMPTIONS` - When we're merging entity trees that
come from different ingest views, it technically matters which order we
merge those trees in because if two ingest views hydrate the exact same
field, we'll pick the second value over the first (e.g. if two different
person views hydrate the `gender` enum). The most important thing is
that this ordering is deterministic, but for the sake of validation, we
added logic that allowed us to preserve the ordering that was present
pre-IID. Now that IID has shipped, we can just order ingest views
alphabetically to ensure determinism.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 3d32ef59a731bf2d8f2f91aa778ca7109c61f7d6