[IID Cleanup] Replace Ingest Rerun issue template with a raw data reimport template (Recidiviz/recidiviz-data#27832)
## Description of the change

We don't need to file tasks for ingest reruns anymore post-IID, but we
may still want to file tasks for raw data reimport work. This PR
replaces the ingest rerun issue template with one that is specifically
for reimporting raw data in secondary.

TODO before merging:
- [x] Create the new `Subject: Ingest Raw Data Re-import` label
- [x] Make sure all legacy ingest rerun tasks are closed

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930.
Closes Recidiviz/recidiviz-data#27832.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 37e2154a8fb77ab02b4546ec23c90fc95e849c08