[US_OR] Ingesting Program Assignments (Recidiviz/recidiviz-data#27257)
## Description of the change

Adding in programs people on supervision have been referred to, are in,
or completed. Conditions attached to supervision sentences should have
the code for the program (will match after the - for the program codes
here since before the - is the county generally, sometimes there is no
dash and it's just the program).

The other types of StateProgramStatusParticipation do not apply here. It
would be good to have one for has not yet started since it seems
relevant in OR that people are referred but have not started and this is
seen as not complying with supervision case plan. Could add to pending
if that makes it easier down stream.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#27821

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4e54dd69639d2fdaef5e7ef2e005fbeb7c347a40