[Workflows] Allow requests for configs for all workflows states (Recidiviz/recidiviz-data#27988)
## Description of the change

The auth layer on the workflows routes was rejecting any state that did
not have an external request route.

This PR updates the workflows routes to allow any enabled workflows
state. In addition, the opportunities endpoint will now return `{}`
instead of throwing an error when an unconfigured state is called.

I cut a ticket to move the external routes into their own blueprint in
https://github.com/Recidiviz/recidiviz-data/issues/27987

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Follow-up to Recidiviz/recidiviz-data#27640

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5077f819a3b37b21b66c007dd0f031156e3906fe