[IID Cleanup] Remove unnecessary state-segmented refresh functionality (Recidiviz/recidiviz-data#27847)
## Description of the change

The OPERATIONS schema is currently refreshed on a per-state basis. This
means that when BQ refresh was paused for a STATE via
`cloud_sql_to_bq_config.yaml` , we also wouldn't refresh operations data
for that state (the idea was that the operations data in BQ should agree
with the data in the state dataset). Now that BQ refresh doesn't apply
to the STATE schema anymore, it doesn't make sense to refresh the
OPERATIONS db this way anymore.

Tested by running this load to a sandbox and confirmed data from all
states gets loaded:
```
    python -m recidiviz.tools.postgres.load_postgres_to_bq_sandbox \
        --project_id recidiviz-staging \
        --sandbox_dataset_prefix ageiduschek \
        --schema OPERATIONS
```

TODO before merging:
- [x] Add a post-deploy task to the prod deploy checklist to delete
`cloud_sql_to_bq_config.yaml` out of both prod and staging GCS
- [x] Audit ingest docs for references to cloud_sql_to_bq_config.yaml
and delete them

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#20930

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e3a875ac57be5915a4c990b29ec29f87b2fdf02a