[Platformization] Query for most relevant active configs (Recidiviz/recidiviz-data#28014)
## Description of the change

This PR builds out the `WorkflowsQuerier` to support queries on the
configs as well as filtering based on feature variants and allowed
systems.

This PR is broken into four commits:
1. Add filtering for opportunities based on active feature variants and
allowed configs
2. Pull in fields from the configs defined in
`recidiviz/calculator/query/state/views/reference/workflows_opportunity_configs.py`
into the query response (some or all of these will eventually move to
the db)
3. Add support for querying the opportunity config table for all active
configs for a type
4. Add a method to return the most relevant config for each provided
type given active feature variants.

Much of the changes are tests, so this should hopefully not be as scary
as the line count suggests.

After this is approved, it should not be much work to hook these
functions up to an endpoint to return all active configs for a user.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#27834

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 6dad9a9eaf88f992a7654794c0f08c17d1ef141e